Move services CLI helpers into lib/services
Drops the underscore-prefixed _reconcile.ts and _resolvePort.ts from
src/commands/services and relocates them to src/lib/services as
reconcileLogger.ts and resolvePort.ts. The services start/stop/restart/
teardown commands import them from the new location.
